Update RedisCache workbook locations & CODEOWNERS
* Copy the RedisCache workbooks to new locations to match other
  workbooks.
* Update the gallery entries to use the new locations.
* Add the azure-cache-for-redis team as a code owner for both the old
  and new locations.
* Update the `id` and `name` properties in the Azure Cache for Redis
  gallery files to more closely match the naming styles that other
  services use for those properties.

Commits f22208b7, 6631860d, and cef04530 moved the workbooks for
ADXCluster, EventHub, and ServiceBus out of the Workbooks/Insights and
Workbooks/Resource Insights directories into per-team directories. The
RedisCache workbooks had not yet been moved from those old directories
to a per-team directory, so they can now be copied into a per-team
directory.

One reason to copy the workbooks into the new directory instead of
moving them is that the Azure Monitor menu and Azure Cache for Redis
resource menu in the Azure portal currently still link to the old
locations. Later, after the new workbook locations become available,
those menus can be updated, and then the content in the old workbook
locations can be replaced with links to the new workbook locations.